Share the cancel and label-upload dialogs between both panels
Both existed twice, once per panel, with the same fields in a different
order and different words for the same things. They now render one shared
form each, with the parts that genuinely differ passed in: the operator's
refund switch, which the seller branch cannot offer because its endpoint
accepts no refund argument at all.

Two gaps closed by the merge. The seller was uploading shipping labels to
public storage while the model attaches them to the private service, so the
attachment pointed at the wrong place for a document carrying a customer's
address. And the seller could submit while a file was still uploading, which
silently recorded a label with nothing behind it — the signed id does not
exist until the upload resolves.

The operator also gains the customer notification its own cancel endpoint
already accepted but its form never offered.
