+++ packages/dashboard-ui/src/index.ts
@@ -46,6 +46,8 @@ export * from './spree/fulfill-items-form'
 export * from './spree/fulfillment-edit-dialog'
 export * from './spree/fulfillment-item-list'
 export * from './spree/fulfillment-panel'
+export * from './spree/label-upload-dialog'
+export * from './spree/order-cancel-dialog'
 export * from './spree/post-sale-create-dialogs'
 export * from './spree/post-sale-fields'
 export * from './spree/return-dialogs'
