+++ packages/dashboard/src/locales/ar.json
@@ -2661,6 +2661,7 @@
           "no_reason": "لم يُذكر سبب",
           "note_label": "ملاحظة",
           "note_placeholder": "يرى هذه الملاحظة الموظفون فقط",
+          "notify_customer": "إرسال بريد إلكتروني إلى العميل",
           "reason_label": "السبب",
           "refund_payments": "رد المبلغ إلى العميل",
           "refund_payments_help": "تُعاد الأموال التي جرى تحصيلها. أما التفويض غير المُحصَّل فيُحرَّر في كل الأحوال."
