+++ packages/dashboard/src/locales/fr.json
@@ -2643,6 +2643,7 @@
           "no_reason": "Aucun motif indiqué",
           "note_label": "Note",
           "note_placeholder": "Cette note n'est visible que par le personnel",
+          "notify_customer": "Envoyer un e-mail au client",
           "reason_label": "Motif",
           "refund_payments": "Rembourser le client",
           "refund_payments_help": "L'argent déjà encaissé est restitué. Une autorisation non utilisée est libérée dans tous les cas."
