+++ packages/dashboard/src/locales/zh-CN.json
@@ -2636,6 +2636,7 @@
           "no_reason": "未填写原因",
           "note_label": "备注",
           "note_placeholder": "此备注仅员工可见",
+          "notify_customer": "通过电子邮件通知客户",
           "reason_label": "原因",
           "refund_payments": "向客户退款",
           "refund_payments_help": "已扣取的款项将退回。未使用的授权无论如何都会释放。"
