+++ packages/dashboard/src/schemas/order.ts
@@ -34,22 +34,6 @@ export type NewOrderFormValues = z.infer<typeof newOrderFormSchema>
  * Cancelling an order. The reason is an id from the store's own cancellation
  * vocabulary, and an empty string means the merchant gave none.
  */
-export const cancelOrderFormSchema = z.object({
-  cancel_reason_id: z.string(),
-  cancel_note: z.string(),
-  refund_payments: z.boolean(),
-})
-
-export type CancelOrderFormValues = z.infer<typeof cancelOrderFormSchema>
-
-export const CANCEL_ORDER_DEFAULTS: CancelOrderFormValues = {
-  cancel_reason_id: '',
-  cancel_note: '',
-  // Off by default, like the API: releasing the gateway's hold is automatic,
-  // but handing back money already taken is the merchant's call.
-  refund_payments: false,
-}
-
 export const NEW_ORDER_DEFAULTS: NewOrderFormValues = {
   email: '',
   customer_note: '',
